+++ source/core/base.js
@@ -244,6 +244,7 @@ class Base {
           tag_e: opt['tag_e'],
           encode: this.__opts__['encode'],
           ignoreHTTPS: (this.__opts__['otherConf'] || {})['ignore-https'] === 1,
+          useMultipart: (this.__opts__['otherConf'] || {})['use-multipart'] === 1,
           timeout: parseInt((this.__opts__['otherConf'] || {})['request-timeout']),
           headers: (this.__opts__['httpConf'] || {})['headers'] || {},
           body: (this.__opts__['httpConf'] || {})['body'] || {}
@@ -285,6 +286,8 @@ class Base {
           data: opt['data'],
           tag_s: opt['tag_s'],
           tag_e: opt['tag_e'],
+          ignoreHTTPS: (this.__opts__['otherConf'] || {})['ignore-https'] === 1,
+          useMultipart: (this.__opts__['otherConf'] || {})['use-multipart'] === 1,
           encode: this.__opts__['encode']
         });
     })